code refactor to make it more responsive
