+++ src/components/ModelFoot.jsx
@@ -1,0 +1,10 @@
+import { Box } from '@mui/material'
+import React from 'react'
+
+export default function ModelFoot() {
+  return (
+    <Box >
+        Hello world
+    </Box>
+  )
+}
