+++ src/components/ModelSide.jsx
@@ -1,0 +1,10 @@
+import { Box } from '@mui/material'
+import React from 'react'
+
+export default function ModelSide() {
+  return (
+    <Box>
+        Hello world
+    </Box>
+  )
+}
