+++ src/index.css
@@ -1,14 +1,14 @@
-/* index.css */
 * {
   box-sizing: border-box;
 }
 
 body {
   margin: 0;
+  background-color: gainsboro;
 }
 
+/* 
 .container {
-  background-color: gainsboro;
   display: flex;
   align-items: center;
   justify-content: center;
@@ -72,4 +72,4 @@ body {
   .text-container {
     left: 0;
   }
-}
+} */
