+++ src/models/pants/Model.jsx
@@ -8,13 +8,15 @@ export function Model(props) {
   // Use useFrame to perform updates on each frame
   useFrame(({ clock }) => {
     const elapsedTime = clock.getElapsedTime();
-    const rotationSpeed = 0.5; 
+    const rotationSpeed = 0.5;
 
     // Rotate 360 degrees over time
     model.scene.rotation.y = elapsedTime * rotationSpeed;
   });
 
-  return <mesh position={[1, -5, 4]}>
-              <primitive object={model.scene}/>
-          </mesh>
+  return (
+    <mesh position={[1, -5, 4]}>
+      <primitive object={model.scene} />
+    </mesh>
+  );
 }
