+++ src/models/pants/Pant.jsx
@@ -1,21 +1,26 @@
 import { Canvas } from "@react-three/fiber";
-import React, { Component } from "react"; 
+import React, { Component } from "react";
 import { Environment, OrbitControls } from "@react-three/drei";
 import { Model } from "./Model";
+import { Box } from "@mui/material";
 
 export class Pants extends Component {
   render() {
-    return ( 
-      <div className="canva">
-      <Canvas  camera={{ position: [2, 0, 15] }}>
-        <Model />
-        <Environment preset="sunset" />
-        <OrbitControls   
-          enableZoom={true}
-          zoomSpeed={0.5} 
-          />
-      </Canvas>
-    </div> 
+    return (
+      <Box
+        sx={{
+          height: "70vh",
+          overflow: "hidden",
+          margin: 0,
+          padding: 0,
+        }}
+      >
+        <Canvas camera={{ position: [2, 0, 15] }}>
+          <Model />
+          <Environment preset="sunset" />
+          <OrbitControls enableZoom={true} zoomSpeed={0.5} />
+        </Canvas>
+      </Box>
     );
   }
 }
